make_symmetric_task fix partial update, warning fixes, matdes devel updates
Currently the added make_symmetric_task machinery doesn't have much effect... only the GreedyOpt stuff is actually "fixed".
I will be making another update soon with a rosetta_scripts interface.

I had no test changes, though oop_dock_design oop_create and features do not currently run.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@48920 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 0e9185a9d7a0c14c5e5d4b1e51541905955f6743

+++ rosetta_source/src/apps/pilot/will/test_ikrs.cc
@@ -1319,7 +1319,7 @@ int main (int argc, char *argv[]) {
       { wp=pose; lg=ctp; }
       wp.replace_residue(ir,lys.residue(1),true);
       for(Size jr = 3; jr <= pose.n_residue()-2; ++jr) {
-        if(ir==jr) continue;
+        if(ir==(int)jr) continue;
         if(sasa[jr] > 0.1) continue;
         //if( (ir*jr+jr+3*ir+999999999)%10!=0 ) continue;
         wp.replace_residue(jr,asp.residue(1),true);
@@ -1351,7 +1351,7 @@ int main (int argc, char *argv[]) {
       { wp=pose; lg=ctp; }
       wp.replace_residue(ir,arg.residue(1),true);
       for(Size jr = 3; jr <= pose.n_residue()-2; ++jr) {
-        if(ir==jr) continue;
+        if(ir==(int)jr) continue;
         if(sasa[jr] > 0.1) continue;
         //if( (ir*jr+jr+3*ir+999999999)%5!=0 ) continue;
         wp.replace_residue(jr,asp.residue(1),true);
